+++ src/components/index.js
@@ -1,5 +1,6 @@
 export { default as InstanceAccordion } from './InstanceAccordion';
 export { default as HoldingAccordion } from './HoldingAccordion';
+export { default as ItemAccordion } from './ItemAccordion';
 export { default as PrimaryToggleButton } from './PrimaryToggleButton';
 export { default as ContributorFields } from './ContributorFields';
-
+export { default as CirculationNoteFields } from './CirculationNoteFields';