+++ test/bigtest/network/serializers/material-type.js
@@ -1,0 +1,4 @@
+
+import OkapiSerializer from './okapi';
+
+export default OkapiSerializer.extend({ name: 'mtypes' });